% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{api_cget}
\alias{api_cget}
\title{Conditional GET}
\usage{
api_cget(api, path, etag_fn)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{path}{the path to short circuit.}

\item{etag_fn}{a function that takes no arguments and returns a single value used to derive the ETag.}
}
\description{
Creates a conditional GET handler for a specific
HTTP path, using a supplied etag function
for the current representation. If the server's ETag and
client's 'If-None-Match' headers match, a \verb{304 Not Modified} response
is sent, short-circuiting downstream handlers. This reduces
the need to recompute responses for paths where the data
has not changed since the last client request.
}
\details{
See \url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Conditional_requests} for more information.
}
\section{Annotation Reference}{
The etag function is specified by \verb{@etag fn} where \code{fn} is the function name.
Functions can also be defined in-line
like \verb{@etag \\() x + 1}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{increment_x <- \\() \{
    x <<- x + 1L
    later::later(increment_x, delay = 10L)
\}
x <- 1L
increment_x()
#* @get /
#* @etag \\() x
function() \{
 x
\}
}\if{html}{\out{</div>}}
}

