% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{api_error_pages}
\alias{api_error_pages}
\title{Freshwater Error Pages}
\usage{
api_error_pages(api, server_error = NULL, not_found = NULL)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{server_error}{TODO}

\item{not_found}{TODO}
}
\description{
todo
}
