% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{api_freshwater}
\alias{api_freshwater}
\title{Freshwater defaults for plumber2 APIs}
\usage{
api_freshwater(api, csrf = TRUE, error_pages = TRUE, ...)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{csrf}{whether to enable CSRF protection}

\item{error_pages}{whether to enable error pages}

\item{...}{args passed to either \code{\link[=api_csrf]{api_csrf()}} or
\code{\link[=api_error_pages]{api_error_pages()}}}
}
\description{
Installs freshwater defaults onto a plumber2 API.

This is a convenience wrapper:
\itemize{
\item Registers freshwater's HTML serialiser
\item Installs freshwater request context
\item Optionally enables CSRF protection
\item Optionally installs HTML error page handlers
}

Arguments in \code{...} are selectively forwarded to
\code{\link[=api_csrf]{api_csrf()}} and \code{\link[=api_error_pages]{api_error_pages()}} based on
matching formal parameters.
}
\seealso{
\link{api_csrf}, \link{api_error_pages}, \link{register_html_serialiser}
}
