% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{document}
\alias{document}
\title{HTML Document Root}
\usage{
document(...)
}
\arguments{
\item{...}{user-supplied content}
}
\value{
An \code{htmltools::tagList}, consisting of a doctype declaration, an \if{html}{\out{<html>}} tag, and
user-supplied content.
}
\description{
Constructs a full HTML document. It does not modify or validate its contents.

\code{document()} is used for full-page responses, and should not be used for partials, fragments
or nested templates.
}
\examples{
document(
     htmltools::tags$head(
         htmltools::tags$title("Home")
     ),
     htmltools::tags$body(
         htmltools::tags$h1("Hello")
     )
)
}
\seealso{
\link{template}, \link{fragment}
}
