% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_hooks.R
\name{enhook_routes}
\alias{enhook_routes}
\title{Handler Hooks}
\usage{
enhook_routes(api, hooks, .where = c("append", "prepend"))
}
\description{
Sometimes you need to get the results of a \link[plumber2:plumber2-package]{plumber2::plumber2} handler \emph{before}
it is processed by the server. Or you need to guarantee that some
checks occur before any further routing.
}
