% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{api_error_pages}
\alias{api_error_pages}
\title{Freshwater Error Pages}
\usage{
api_error_pages(
  api,
  handlers = NULL,
  debug = plumber2::get_opts("fw_debug", default = interactive())
)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{handlers}{optional list of named error templates. Supported keys
are: "403", "404", "500". If omitted, freshwater installs default templates.}

\item{debug}{whether the \strong{500} error template should render error
messages and stack traces. Defaults to the \code{fw_debug} plumber2 option,
and falls back to \code{interactive()}.}
}
\description{
Adds request/error hooks to a \link[plumber2:plumber2-package]{plumber2::plumber2} API so that
freshwater can render friendly \strong{HTML error pages} for:
\itemize{
\item {\strong{403 Forbidden}} responses
\item {\strong{404 Not Found}} responses
\item {\strong{500 Internal Server Error}} conditions
}
}
\details{
Custom error page templates can be supplied via the
\code{handler} parameter. These should be freshwater templates
created via \code{\link[=template]{template()}}, and should match the call signatures of
the default error templates. See \link{freshwater_error_templates} for
the relevant template signatures required.

This middleware installs freshwater request context.
}
\examples{
#* @plumber
function(api) {
 api |>
     api_error_pages(debug = TRUE)
}

}
\seealso{
\link{freshwater_error_templates}, \link{enhook_routes}
}
