% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{freshwater_error_templates}
\alias{freshwater_error_templates}
\alias{default_error_500_template}
\alias{default_error_404_template}
\alias{default_error_403_template}
\title{Error Page Templates}
\usage{
default_error_500_template(
  error = NULL,
  request = NULL,
  is_debug = FALSE,
  ...,
  fragment = NULL
)

default_error_404_template(error = NULL, request = NULL, ..., fragment = NULL)

default_error_403_template(error = NULL, request = NULL, ..., fragment = NULL)
}
\arguments{
\item{error}{the error condition signaled by an error in the server's route handler}

\item{request}{the \link[reqres:Request]{reqres::Request} request object the handler is responding to}

\item{is_debug}{whether to provide the stack trace and error message to the
web client. Although useful during development, it is heavily recommended to
set as \code{FALSE} in production as it can leak sensitive information.}

\item{...}{unused}

\item{fragment}{unused}
}
\description{
freshwater provides a number of default views
that are served to HTML clients in the event of common HTTP
error codes.
}
\seealso{
\link{api_error_pages}, \link{template}
}
