% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cache.R
\name{get_cache_backend}
\alias{get_cache_backend}
\title{Get freshwater's current cache backend}
\usage{
get_cache_backend()
}
\value{
A cache backend object (typically from \code{cachem}), or \code{NULL} if the
cache has not yet been initialised.
}
\description{
Returns the cache backend currently used by \code{\link[=cache]{cache()}}.
}
\examples{
get_cache_backend()

}
\seealso{
\code{\link[=set_cache_backend]{set_cache_backend()}}, \code{\link[=cache]{cache()}}
}
