% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_hooks.R
\name{api_hooks}
\alias{api_hooks}
\alias{hook}
\title{Route handler hooks}
\usage{
api_hooks(api, hooks, .where = c("append", "prepend"))

hook(id = NULL, fn)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{hooks}{a single hook or list of hooks that take the signature
\code{fn(api, args, next_call)}, where \code{args}
is the list of handler arguments. If a hook returns without calling \code{next_call()}, the
remaining hooks and user handler are skipped, and the return value becomes the handler result.}

\item{.where}{whether the hooks should be appended or prepended to the
list of installed hooks}

\item{id}{id of the hook}

\item{fn}{function with signature \code{fn(api, args, next_call)}}
}
\description{
Add middleware-style hooks to \emph{all
existing user handlers} in a \link[plumber2:plumber2-package]{plumber2::plumber2} API.
Hooks execute in a deterministic order prior to the user handler,
and can mutate/intercept requests and responses,
as well as short-circuit handlers. This allows for middleware-type
behaviour without registering additional routes.
}
\details{
System hooks are installed in the following order:

\itemize{
\item \code{freshwater::context}
\item \code{freshwater::error_pages}
\item \code{freshwater::csrf}
\item \code{freshwater::csrf_context}
}
\subsection{Control Flow}{

Hooks control whether subsequent hooks should execute.
Concretely:
\itemize{
\item To continue to the next hook (and eventual user handler), call \code{next_call()}.
\item To short-circuit the chain, return a value without calling \code{next_call()}.
\item To bubble up to plumber2 routing control flow, return either \code{plumber2::Next}
or \code{plumber2::Break} (and don't call \code{next_call()}).
}

Hooks can also wrap later hooks and the user handler by calling \code{next_call()},
and then doing work after.
}

\subsection{Hook Installation}{
\itemize{
\item Routing is managed by \link[plumber2:plumber2-package]{plumber2::plumber2} -- this function does not change
routing precedence. Within the handler itself, however, hooks run in the order
they are installed.
\item The function is idempotent (with respect to either a computed hash of the hook or a
provided id), and only new hooks will be installed.
}
}

\subsection{Asynchronous Routes}{

When using asynchronous routes via \code{async=TRUE}
programmatically, or via \verb{@async}, hooks are attached to
the \code{then} handlers, rather than main handler itself. This is because
\code{request}, \code{response}, and \code{server} arguments
are not available to the main async
handler, and hooks depend on the
full handler signature being available.
}
}
\examples{
api <- plumber2::api() |>
  plumber2::api_get(path = "/", function() {
    "Foo"
  })

log_hook <- hook(
  "logger",
  function(api, args, next_call) {
    msg <- sprintf(
      "[\%s] \%s \%s",
      format(Sys.time(), "\%H:\%M:\%S"),
      args$request$method,
      args$request$path
    )
    print(msg)
    next_call()
  }
)

timer_hook <- hook("timer", function(api, args, next_call) {
  t0 <- Sys.time()
  out <- next_call()
  print(sprintf("time: \%s", Sys.time() - t0))
  out
})
api <- api_hooks(api, hooks = list(log_hook, timer_hook))

}
