% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{map_tags}
\alias{map_tags}
\title{Apply template function to each element of a vector}
\usage{
map_tags(.x, .f, .empty = NULL)
}
\arguments{
\item{.x}{list or atomic vector}

\item{.f}{a function that takes a single argument returns a character
vector, tag, or tagList.}

\item{.empty}{fallback value for falsey elements}
}
\description{
\code{map_tags()} returns a type-safe tag list,
where each element is resolved by applying \code{.f} to each
element of \code{.x}. If an element in \code{.x} is a falsey value
(i.e. \code{NA}, \code{NaN}, \code{FALSE}, or \code{NULL}), the fallback value
from \code{.empty} is used.

Additional arguments should be passed
with an anonymous function.
}
\details{
Element values are evaluated prior to returning the tag list:
\itemize{
\item NULL values are removed from the final tag list.
The return length of \code{map_tags()} is therefore less
than or equal to the length of \code{.x}
\item All return values must be either a "shiny.tag",
"shiny.tag.list", or "character" vector. An
error is raised if an unexpected value is encountered.
}
}
\examples{
tpl <- template(x, {p(x)})
map_tags(seq(5L), tpl)

# falsey values are removed
map_tags(c(TRUE, FALSE, TRUE), tpl)
}
\seealso{
\link{template}, \link[base:lapply]{base::lapply}, \link[htmltools:tagList]{htmltools::tagList}
}
