% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\name{mycols}
\alias{mycols}
\title{mycols}
\usage{
mycols(...)
}
\description{
Extract the hex codes from a list of colours.
}
\examples{
mycols("green")
mycols("green", "blue")
normalHist(u,dens=TRUE)
ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
geom_point(color = mycols("red"),
           size = 1.5, alpha =1)+
  theme_test()

}
\author{
Carlos Cano-Barbacil
}
