% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect.R
\name{redirect}
\alias{redirect}
\title{Redirect to another resource}
\usage{
redirect(response, location, after = NULL)
}
\arguments{
\item{response}{\link[reqres:Response]{reqres::Response} object}

\item{location}{path or url to redirect to}

\item{after}{optional number of seconds to wait before redirection}
}
\value{
\itemize{
\item \link[plumber2:Next]{plumber2::Break} when issuing an immediate redirect.
\item \link[plumber2:Next]{plumber2::Next} when issuing a delayed navigation.
}
}
\description{
If \code{after} is NULL, sends a 303 response and halts request processing.
Client is redirected to the given location. This is commonly used in
Post/Redirect/GET (PRG) setups to redirect clients to a new page following
form submissions.
}
\details{
If after is a numeric, a "Refresh" header is attached to the response,
instructing the browser to navigate to \code{location} after the specified number
of seconds.

The delayed redirect uses the non-standard "Refresh" HTTP header which is
widely supported by browsers but is not part of the official HTTP specification.
It should not be relied on for API & non-browser clients.

See also:
\itemize{
\item \url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Redirections}
\item \url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Refresh}
}
}
\examples{
# Immediate redirect (PRG pattern)
#* @get /
function(response) {
    print("Hello!")
    redirect(response, "/foo")
}

# Delayed redirect after rendering content
#* @get /count/<n>
function(n, response) {
  redirect(response, "/", after = 1)
  paste("n =", n)
}
}
