% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{register_async_evaluator}
\alias{register_async_evaluator}
\title{Register context-safe async evaluator}
\usage{
register_async_evaluator(force = FALSE, set_default = TRUE)
}
\arguments{
\item{force}{whether to register the evaluator regardless of if it has been
registered already}

\item{set_default}{whether to set the default async evaluator to \code{"freshwater"}}
}
\description{
Registers an asynchronous evaluator for routes, allowing freshwater context to
be propagated to \link[mirai:mirai]{mirai::mirai} workers. This means that contextful helpers such as \link{current_method},
\link{current_path}, and \link{current_query} (among others) will work in async routes.
}
\details{
Context is not inherently portable across asynchronous contexts, this function
creates a portable snapshopt of the current context that is passed to a mirai worker.

Hooks are \emph{not} applied to the async route, but may be provided to any associated
\code{then} handlers. If error pages are installed on the main process, errors from
the worker will be appropriately converted into freshwater error pages.

Requires the \link[promises:promises-package]{promises::promises}, \link[mirai:mirai]{mirai::mirai}, and \link[mori:mori-package]{mori::mori} packages.
}
\examples{
register_async_evaluator()
#* @async
#* @get /async
function() {
 current_path()
}
}
\seealso{
\link{api_freshwater}, \link{api_error_pages}, \link{api_hooks}, \link[mirai:mirai]{mirai::mirai}, \link{current_context}
}
