% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisers.R
\name{register_html_serialiser}
\alias{register_html_serialiser}
\title{Register HTML Serialiser}
\usage{
register_html_serialiser(force = FALSE)
}
\description{
Registers an HTML serialiser for \link[plumber2:plumber2-package]{plumber2::plumber2}
that renders shiny tags and taglists via
\code{\link[htmltools:renderTags]{htmltools::doRenderTags()}}. This is preferrable
over \code{\link[htmltools:renderTags]{htmltools::renderTags()}} as often we want to
be able to emit head tags which \link[htmltools:htmltools-package]{htmltools::htmltools} attempts
to hide for shiny usecases.
}
\details{
The freshwater serialiser safely falls back
to the default plumber2 implementation for other
classes of inputs.
}
