% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\name{scale_color_freshwater}
\alias{scale_color_freshwater}
\title{scale_color_freshwater}
\usage{
scale_color_freshwater(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Colour your ggplot2 graphs with FreshWater palettes.
}
\author{
Carlos Cano-Barbacil
}
