% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{target}
\alias{target}
\title{Title}
\usage{
target(tpl, ..., .part = NULL)
}
\arguments{
\item{tpl}{TODO}

\item{...}{TODO}

\item{part}{TODO}
}
\description{
Title
}
\examples{
card <- template(
    user,
    .id = function(user) sprintf("user-\%s", user$id),
    {
        div(user$name)
    }
)
target(card, list(id = 1234L))

user_table <- template(users = list(), .id = "my-table", {
    table(
        thead(
            .part = "header",
            tr(th("Name"))
        ),
        tbody(
            .part = "body",
            lapply(users, \(user) tr(td(user)))
        ),
        tfoot(
            .part = "footer",
            tr(td(sprintf("There are '\%s' users.", length(users))))
        )
   )
})
}
