% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{targets}
\alias{targets}
\title{Combine multiple target selectors}
\usage{
targets(...)
}
\arguments{
\item{...}{templates or character vectors}
}
\description{
\code{targets()} returns a comma-separated string combining multiple
\code{\link[=target]{target()}} calls. If a template is supplied, \code{target()} is called on it,
otherwise the value is coerced to character and used as-is.
}
\examples{
tpl <- template(.id = "foo", {})
tpl2 <- template(x, .id = function(x) x, {})
targets(
    tpl,
   target(tpl2, x = "bar")
)
}
\seealso{
\link{target}
}
