% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cache.R
\name{cache}
\alias{cache}
\alias{clear_cache}
\alias{invalidate_cache}
\alias{invalidate_cache_here}
\title{Clear the cache of all memoised templates}
\usage{
cache(name, vary = NULL, ...)

clear_cache()

invalidate_cache(tpl, name, vary = NULL, fragment = NULL)

invalidate_cache_here(name, vary = NULL, fragment = NULL)
}
\arguments{
\item{name}{unique name for the cached partial template}

\item{vary}{values that should change when the cached output should change. This is used to construct the cache key.}

\item{...}{tag content to render and cache}

\item{tpl}{a template function created by \code{\link[=template]{template()}}.}

\item{fragment}{optional fragment name for targetting cached fragments}
}
\description{
Clear the cache of all memoised templates
}
\examples{
# Caching
nav <- template(user, {
  div(
    cache(
      "nav",
      vary = user$id,
      ul(
        li("Home"),
        li("Profile"),
        if (user$is_admin) li("Admin")
      )
    )
  )
})
nav(list(id = 1, is_admin = TRUE))

# Nested Caches
dashboard <- template(page = list(), stats = list(), recent = list(), {
    cache(
        name = "page",
        vary = page$updated_at,
        div(
            h1("Dashboard"),
            cache(
                name = "stats",
                 vary = stats$updated_at,
                div(p(stats$count))
            ),
            cache(
                name = "recent",
                vary = recent$updated_at,
                div(recent)
            )
        )
    )
})
dashboard()

# Invalidate the current cache
# during rendering

page <- template(user, {
  div(
    cache(
      name = "content",
      vary = user$id,
      {
        if (user$refresh) {
          invalidate_cache_here(
            name = "content",
            vary = user$id
          )
        }
        p("Hello ", user$id)
      }
    )
  )
})

page(list(id = 1, refresh = FALSE))

page(list(id = 1, refresh = TRUE))

}
\seealso{
\link{template}
}
