% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating.R
\name{template}
\alias{template}
\title{Create a reusable HTML template with optional fragments}
\usage{
template(...)
}
\arguments{
\item{...}{content}
}
\description{
Evaluates code under the \code{htmltools::withTags()} environment,
creating a closure that outputs either the template or
specified fragment.
}
\examples{
page_main <- template(
    div(
        h1("Dashboard"),
        fragment(p("Welcome back"), name="content"),
        small("2026")
    )
)
page_main(fragment="content")

}
