% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{api_freshwater}
\alias{api_freshwater}
\title{Freshwater defaults for plumber2 APIs}
\usage{
api_freshwater(api, csrf = TRUE, error_pages = TRUE, ...)
}
\arguments{
\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object.}

\item{csrf}{whether to enable CSRF protection}

\item{error_pages}{whether to enable error pages}

\item{...}{args passed to either \code{\link[=api_csrf]{api_csrf()}} or
\code{\link[=api_error_pages]{api_error_pages()}}}
}
\description{
This is a convenience wrapper around
\code{\link[=register_html_serialiser]{register_html_serialiser()}}, \code{\link[=api_csrf]{api_csrf()}}, and
\code{\link[=api_error_pages]{api_error_pages()}}.
}
