% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{csrf_token}
\alias{csrf_token}
\title{CSRF Token}
\usage{
csrf_token()
}
\description{
\code{csrf_token()} returns the current CSRF token string for
the active request when used within a \code{\link[=template]{template()}}. Calling it
outside of a \code{template()} context will result in an error.

In most cases, CSRF tokens are inserted automatically
for standard form helpers.  Intended for custom forms / custom token
placement (meta tags, JS fetch, etc).
}
\examples{
page <- template({
    html(
        head(
            meta(name = "csrf-token", content = csrf_token())
        ),
        body(
            div("App content")
        )
    )
})
page()
}
\seealso{
\link{api_csrf}
}
