% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{csrf_token}
\alias{csrf_token}
\title{CSRF Token}
\usage{
csrf_token()
}
\description{
\code{csrf_token()} returns the current CSRF token string for
the active request when used within a \code{\link[=template]{template()}}.
Intended for custom forms / custom token placement
(meta tags, JS fetch, etc).

Do not call the \code{csrf_token()} function directly, it is a stub.
}
\seealso{
\link{api_csrf}
}
