% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{endpoints}
\alias{endpoints}
\title{Reverse Routing}
\usage{
endpoints(route = NULL, api = NULL, refresh = FALSE)
}
\arguments{
\item{route}{the route group to retrieve endpoints from; typically defined either via the file name, routeName or route in plumber2. If NULL, will return all endpoints for all routes.}

\item{api}{a \link[plumber2:plumber2-package]{plumber2::plumber2} api object. If NULL, context is used to find the api.}

\item{refresh}{force refresh the registered routes. Useful if you have added routes after calling \code{endpoints()}.}
}
\value{
If \code{route} is \code{NULL}, returns a list of route groups
and their endpoints. Otherwise returns a list of a
route's endpoint accessors.
}
\description{
Access generated endpoint URL helpers.
}
\details{
Alias rules:
\itemize{
\item "/" endpoints become "index"
\item GET endpoints are accessed directly, like index()
\item non-GET endpoints require an accessor, like index$delete()
\item path parameters are removed from the alias and supplied as function args
}

For example:
\itemize{
\item \verb{GET /} -> \code{index()}
\item \verb{POST /} -> \code{index$post()}
\item \code{GET /my/filter} -> \code{my_filter()}
\item \verb{GET /users/:id} -> \code{users(id = 1)}
\item \verb{DELETE /users/:id} -> \code{users$delete(id = 1)}
}
}
\examples{
#* @plumber
function(api) {
   api |>
     api_freshwater()
}

#* @get /
#* @serializer html
#* @routeName user
function() {
     endpoints("user")$index()
}
}
