% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\name{freshwater_pal}
\alias{freshwater_pal}
\title{freshwater_pal}
\usage{
freshwater_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose your favourite palette: "Tajo", "Jarama", "Manzanares", "Culebro", "Aphanius", "Lepomis" or "Cyprinus"}

\item{reverse}{TRUE/FALSE}
}
\description{
This function allows to get a pallete by name from the list ("Tajo" by default).
It also has a boolean condition determining whether to reverse the color order or not.
}
\examples{
freshwater_pal("Tajo")
colors <- freshwater_pal("Tajo")(12)
colors
show_col(colors)

}
\author{
Carlos Cano-Barbacil
}
