% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect.R
\name{redirect}
\alias{redirect}
\title{Redirect to another resource}
\usage{
redirect(response, location)
}
\arguments{
\item{response}{Plumber2 response object}

\item{location}{path or url to redirect to}
}
\description{
Sends a 303 response and halts request processing.
}
\examples{
#* @get /
function(response) {
    print("Hello!")
    redirect(response, "/foo")
}
}
