% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\name{scale_color_freshwater}
\alias{scale_color_freshwater}
\title{scale_color_freshwater}
\usage{
scale_color_freshwater(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose your favourite palette: "Tajo", "Jarama", "Manzanares", "Culebro", "Aphanius", "Lepomis" or "Cyprinus"}

\item{discrete}{TRUE/FALSE}

\item{reverse}{TRUE/FALSE}
}
\description{
Colour your ggplot2 graphs with FreshWater palettes.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 2, alpha = 1) +
  scale_color_freshwater(discrete = FALSE, palette = "Jarama")+
  ggtitle("Jarama")+theme_test()+ theme(legend.position = "none")

}
\author{
Carlos Cano-Barbacil
}
