% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwater.R
\name{scale_fill_freshwater}
\alias{scale_fill_freshwater}
\title{scale_fill_freshwater}
\usage{
scale_fill_freshwater(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{c("Tajo", "Jarama", "Manzanares", "Culebro", "Aphanius", "Lepomis", "Cyprinus")}

\item{discrete}{TRUE/FALSE}

\item{reverse}{TRUE/FALSE}
}
\description{
Colour your ggplot2 graphs with FreshWater palettes.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(shape = 21, size = 2, alpha = 1) +
  scale_fill_freshwater(discrete = FALSE, palette = "Jarama")+
  ggtitle("Jarama")+theme_test()+ theme(legend.position = "none")

}
\author{
Carlos Cano-Barbacil
}
