% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cache.R
\name{set_cache_backend}
\alias{set_cache_backend}
\title{Configure freshwater's cache backend}
\usage{
set_cache_backend(backend, clear = TRUE)
}
\arguments{
\item{backend}{cache backend accepted by \link[memoise:memoise]{memoise::memoise}}

\item{clear}{If \code{TRUE}, clears the current cache before switching backends.}
}
\description{
\code{set_cache_backend()} replaces the cache backend used by \code{\link[=cache]{cache()}}.
}
\details{
This function allows controlling cache persistence (memory vs disk),
eviction policies, and storage limits via the backend object.
}
\seealso{
\code{\link[=cache]{cache()}}, \code{\link[=clear_cache]{clear_cache()}}, \link[memoise:memoise]{memoise::memoise}
}
