% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{target}
\alias{target}
\title{Resolve a template's target selector}
\usage{
target(tpl, ..., .part = NULL)
}
\arguments{
\item{tpl}{a freshwater template}

\item{...}{arguments passed to the template id function}

\item{.part}{whether to select a sub-part}
}
\description{
\code{target()} returns a CSS selector for a template instance. For normal
targets, this is an id selector of the form \verb{#<TEMPLATE_ID>}, where
the id is resolved from the template's id function.
}
\details{
In order to support multi-root template selection,
supplying \code{.part} will return a data attribute selector of the form:
\verb{[data-fw-part="<TEMPLATE_ID>-<PART_NAME>"]}
Part names are automatically scoped against the template's \code{.id}, ensuring
unique data attributes across templates.
}
\examples{
card <- template(
    user,
    .id = function(user) sprintf("user-\%s", user$id),
    {
        div(user$name)
    }
)
target(card, list(id = 1234L))

user_table <- template(users = list(), .id = "my-table", {
    table(
        thead(
            .part = "header",
            tr(th("Name"))
        ),
        tbody(
            .part = "body",
            lapply(users, \(user) tr(td(user)))
        ),
        tfoot(
            .part = "footer",
            tr(td(sprintf("There are '\%s' users.", length(users))))
        )
   )
})
}
