% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating.R
\name{template}
\alias{template}
\alias{fragment}
\title{Create a reusable HTML template}
\usage{
template(..., .envir = parent.frame())

fragment(..., name = NULL)
}
\arguments{
\item{...}{content}

\item{.envir}{the environment in which to evaluate the template}

\item{name}{the name of the fragment}
}
\description{
Evaluates code under the \code{htmltools::withTags()} environment,
creating a closure that outputs either the template or
specified fragment.

The closure can also accept slots, specified as named or bare symbols prior
to the template expression. This allows for passing in other templates or HTML
tags after template creation.
}
\examples{
# Example Fragment Usage
page_main <- template(
    {
        div(
            h1("Dashboard"),
            fragment(p("Welcome back"), name = "content"),
            small("2026")
        )
    }
)

page_main(fragment="content")

# Template slots

details <- template(name, age, {
    nm <- sprintf("Hello, my name is: \%s", name)
    old <- sprintf("I am \%s years old.", age)
    div(
        p(nm),
        p(old)
    )
})

details("Jim", 30)

# Templates and fragments can also be combined

card <- template(
    ttl, foot = NULL, {
        div(
            h2(ttl),
            fragment(div("Card body"), name="body"),
            if (!is.null(foot)) {
                fragment(
                    div(foot),
                    name = "footer"
                )
            }
        )
    }
)
card("Card Title")
card("Card Title", fragment="body")
card("Card Title", "Footer text", fragment = "footer")

}
