% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating.R
\name{template}
\alias{template}
\alias{fragment}
\alias{cache}
\alias{clear_cache}
\title{Create a reusable HTML template}
\usage{
template(..., .envir = parent.frame())

fragment(..., name = NULL)

cache(name, vary = NULL, ...)

clear_cache()
}
\arguments{
\item{...}{template definition. Provide zero or more parameters, followed by a
single braced expression.}

\item{.envir}{the environment in which to evaluate the template}

\item{name}{the name of the fragment}
}
\value{
function of class \code{template} with interface \verb{fn(<declared params>, ..., fragment = NULL)}
}
\description{
\code{template()} is a function factory that captures a template expression and
returns a callable HTML renderer. The expression is evaluated
under the \code{\link[htmltools:withTags]{htmltools::withTags()}} environment, so tag functions such as
\code{div()} or \code{p()} are available.

Templates may define:
\itemize{
\item \strong{parameters}: symbols or named defaults which are used as arguments
to the renderer
\item \strong{content injection}: if the template uses \code{...}, the renderer
passes \code{...}  to the containing HTML nodes
defined in the template.
\item \strong{fragments}: named subtemplates that can be optionally extracted from
the template upon rendering by supplying \code{fragment = "name"}. Fragment names
are required.
}
}
\examples{
# Example Fragment Usage
page_main <- template(
    {
        div(
            h1("Dashboard"),
            fragment(p("Welcome back"), name = "content"),
            small("2026")
        )
    }
)

page_main(fragment="content")

# Template slots

details <- template(name, age, {
    nm <- sprintf("Hello, my name is: \%s", name)
    old <- sprintf("I am \%s years old.", age)
    div(
        p(nm),
        p(old)
    )
})

details("Jim", 30)

# Templates and fragments can also be combined

card <- template(
    ttl, foot = NULL, {
        div(
            h2(ttl),
            fragment(div("Card body"), name="body"),
            if (!is.null(foot)) {
                fragment(
                    div(foot),
                    name = "footer"
                )
            }
        )
    }
)
card("Card Title")
card("Card Title", fragment="body")
card("Card Title", "Footer text", fragment = "footer")

# Dots (content injection)
layout <- template({
    htmltools::tagList(
        head(meta(title = "foo")),
        body(...)
    )
})

layout(htmltools::div("content"))

Caching
nav <- template(user, {
  cache(
    "nav",
    vary = user$id,
    ul(
      li("Home"),
      li("Profile"),
      if (user$is_admin) li("Admin")
    )
  )
})

}
